pyproject.toml
[tool.poetry]
name = "meduzzen-back-end"
version = "0.1.0"
description = ""
authors = ["tarastm952 <tarasmazepa95@ukr.net>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.8"
sqlalchemy = "^2.0.38"
pydantic = {extras = ["email"], version = "^2.10.6"}
uvicorn = "^0.34.0"
pydantic-settings = "^2.8.0"
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
redis = "^4.0"
aioredis = "1.3.0"
pytest = "8.2.0"
pytest-asyncio = "0.24.0"
alembic = "^1.14.1"
psycopg2-binary = "^2.9.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aiosqlite = "^0.21.0"
werkzeug = "^3.1.3"
requests = "^2.32.3"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"


[tool.poetry.group.dev.dependencies]
httpx = "^0.27.0"
black = "^25.1.0"
flake8 = "^7.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
